LPS-113021 Deprecate `liferay-dynamic-select`
In this change, we move `liferay-dynamic-select`'s  functionality to a
new `DynamicSelect` module in `frontend-js-web`, keeping the original
AUI `liferay-dynamic-select` file in case it is been required.

The global `Liferay.DynamicSelect` (now declared in `frontend-js-web`)
has been marked as deprecated because we want to avoid
suggesting the use of globals, and a `@deprecated` annotation
has also been added to the original `liferay-dynamic-select` AUI module in
`frontend-js-aui-web`.

Usage sites have been updated, in JSP files, except for the one found in
`modules/apps/archived`.

To test this change, you will need to deploy the following modules:

- modules/apps/frontend-js/frontend-js-aui-web
- modules/apps/frontend-js/frontend-js-web
- modules/apps/portal-settings/portal-settings-authentication-token-web
- modules/apps/portal-settings/portal-settings-web
- modules/apps/users-admin/users-admin-web
- portal-web - (use `ant deploy`)

One done, navigate to http://localhost:8080, and from the (left) control
menu, select:
- Control Panel > Users > Users and Organizations
- Select the Organizations tab
- Click on the "+" button to create a new organization