LPS-112628 Uses portletResponse to render from ResourceURL as well
This JSP is rendered both in a regular render phase as well as through a
`ResourceCommand` in a polling interval. When the JSP is rendered in the
resource phase, `renderResponse` seems to be null.

The <portlet:namespace /> implementation always retrieves the portlet
namespace through `(PortletResponse)portletResponse.getNamespace()` which
is why this used to work before and seems like a safest approach.